Initial commit
This takes a binary (any binary right now), signs it with a private key
that we have generated (RSA-PSS). After that it compress the uploaded
firmware, the SHA-256 hash and the RSA signature into a zip and send it
back to the caller.

ACS isn't used here yet, since it's unclear how we intend to use the
signed binary/binaries.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>